Permisos y Policies everywhere (+ Updates + Fixs)
|+ Permisos:
|- Se crearon Policy para todos los modelos que lo requerian
|- Los User Jefes ahora son parte del dominio /admin solo con permisos de lectura
|- Los Insumos, Tipos, Formatos y Configuration se pasaron al dominio /admin
|
|+ Proceso:
|- Al Eliminar un Proceso se elimina la carpeta en storage con todas sus fotos
|
|+ Preevaluacion:
|- Las fotos se agruparon en una carpeta /preevaluacion dentro de la carpeta del proceso
|
|+ Cliente:
|- Ahora tiene acceso a la informacion de sus Vehiculos y Procesos con su User
|
|+ Vehiculo:
|- Se agregaron los Procesos asociados a los Vehiculos en la vista Show
|- No se pueden eliminar Vehiculos si ya tienen Procesos asociados
|- Los Clientes tienen CRUD para administrar sus vehiculos
|
|+ Insumo:
|- Se elimino el parametro $foto innecesario del metodo getPhoto
|
|+ Tipos y Formatos:
|- Fix al obtener el Stock de los Insumos en la vista Show
|
|+ Fix Vehiculo:
|- Al agregar un vehiculo, daba error si no se enviaba un Cliente
|
|+ Fix Agendamiento:
|- En los metodos Edit/Update se enviaba Proceso en lugar del Agendamiento
|- Al crear un Angedamiento, se corrigio el nombre del campo que indica si fue atendido inmediatamente o no

+++ app/Http/Controllers/Admin/AgendamientoController.php
@@ -16,7 +16,9 @@ class AgendamientoController extends Controller
      */
     public function index()
     {
-      $agendamientos = Agendamiento::all();
+      $this->authorize('index', Agendamiento::class);
+
+      $agendamientos = Agendamiento::count();
 
       return view('admin.agendamiento.index', compact('agendamientos'));
     }
@@ -60,7 +62,7 @@ class AgendamientoController extends Controller
 
       $agendamiento = new Agendamiento([
                         'taller' => Auth::id(),
-                        'atender' => isset($request->atender),
+                        'inmediatamente' => $request->has('atender'),
                         'fecha' => $request->atender ? date('Y-m-d H:i:s') : $request->fecha,
                       ]);
 
@@ -100,7 +102,8 @@ class AgendamientoController extends Controller
      */
     public function edit(Agendamiento $agendamiento)
     {
-      $this->authorize('update', [Agendamiento::class, $proceso]);
+      $this->authorize('update', $agendamiento);
+
       return view('admin.agendamiento.edit', compact('agendamiento'));
     }
 
@@ -113,7 +116,7 @@ class AgendamientoController extends Controller
      */
     public function update(Request $request, Agendamiento $agendamiento)
     {
-      $this->authorize('update', [Agendamiento::class, $proceso]);
+      $this->authorize('update', $agendamiento);
       $this->validate($request, [
         'fecha' => 'nullable|date',
       ]);
@@ -142,6 +145,8 @@ class AgendamientoController extends Controller
      */
     public function destroy(Agendamiento $agendamiento)
     {
+      $this->authorize('delete', $agendamiento);
+
       if($agendamiento->delete()){
         return redirect()->route('admin.proceso.show', ['proceso' => $agendamiento->proceso_id])->with([
                 'flash_class'   => 'alert-success',
